refactor: move rendering code to render.rs
